src/microservices/events/requirements.txt
Flask==2.3.2
kafka-python==2.0.2